ShadowDOM works w/ "END" keypress to toggle invisibility of main Cube vs the Shadow cube... and "INSERT" key to copy the main cube to the shadowDOM. (ended up just copying all : edges[i] and all : corners[i] .data's over. - good way for now.) The .swap() function is still broken and i've tried to document its parity errors. right now resetFacesCorners() and resetFacesEdges() are affected at least.
Adjusted camera angle to 200 degrees instead of 180 to make the motion less jarring but its not really helping all that much.
Adjust BEIGE color code to make it darker and less confusing with white.
Add some DirToWhiteFace[12][5] directional commands to dictate how pieces rotate and bubble up to the white face.
Add face-specific pieceNum functions that give you the 1-5 position on the face.
Add a single <int> case for findEdgeByPieceNum() and findCornerByPieceNum() instead of the 5 ints/vectors.
Line 900+ In megaminx.cpp has rotateSolveWhiteEdges() and movePieceByRotatingIt() is being actively written.