BREAK page-down - start working on corner algo.
it solves them with R'D'RD, if the piece is directly below where it needs to drop-in at, it goes to the top face. It  just repeats that algo for however many times 1-5 it takes to match up the colors.
adjust GREEN colors to be more different.
BlackCornersNumber2 isnt real, temporary.